// features/components/projects/hooks/use-projects.ts
import {useMutation, useQuery} from 'convex/react'

import {api} from '../../../../convex/_generated/api';
import { Id } from '@/convex/_generated/dataModel';

export const useProjects = ()=>{
    return useQuery(api.projects.get)

}
export const useProject = (projectId: Id<"projects">)=>{
    return useQuery(api.projects.getById, {id: projectId})
}
export const useProjectsPartial = (limit : number)=>{
    return useQuery(api.projects.getPartial, {limit})

}

export const useCreateProject = ()=>{
    return useMutation(api.projects.create).withOptimisticUpdate(
        (localStore,args)=>{
            const existingProjects = localStore.getQuery(api.projects.get);

            if(existingProjects !== undefined){
                const now = Date.now();
                const newProject = {
                    _id : crypto.randomUUID() as Id<"projects">,
                    name : args.name,
                    _creationTime : now,
                    ownerId : 'local-user' ,
                    updatedAt : now,
                } 
                localStore.setQuery(api.projects.get,{}, [newProject, ...existingProjects])
            }
        }
        



    )
}

export const useRenameProject = (projectId :  Id<"projects">)=>{
    return useMutation(api.projects.rename).withOptimisticUpdate(
        (localStore,args)=>{
            const existingProject = localStore.getQuery(api.projects.getById,{
                id : projectId
            });




            if(existingProject !== undefined && existingProject !== null) {
                localStore.setQuery(api.projects.getById,{id: projectId}, {
                    ...existingProject,
                    name : args.name,
                    updatedAt : Date.now(),
                })
            }

             const existingProjects = localStore.getQuery(api.projects.get);

            if(existingProjects !== undefined){

                localStore.setQuery(api.projects.get,{},existingProjects.map((project)=>{
                    return project._id === projectId ? {
                        ...project,
                        name : args.name,
                        updatedAt : Date.now(),
                    } : project
                }))
            }

        }
        
           



    )
}

